x/farmchain/types/messages_location_data_test.go
package types

import (
	"testing"

	"github.com/cosmonaut/farmchain/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateLocationData_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateLocationData
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateLocationData{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateLocationData{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateLocationData_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateLocationData
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateLocationData{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateLocationData{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteLocationData_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteLocationData
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteLocationData{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteLocationData{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
